requirements.txt
torch
torchvision
Pillow
scikit-learn
umap-learn
hdbscan
pyyaml
tqdm
numpy
matplotlib
opencv-python
